src/requirements.txt
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
Flask==2.0.2
Flask-Compress==1.10.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycparser==2.21
pyOpenSSL==21.0.0
python-dotenv==0.19.2
requests==2.26.0
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2
